pyterrier_enterprise/requirements.txt
python-terrier>=0.11.0
tika>=2.6.0
pandas>=2.2.2
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.1
